backend/app/core/security.py
from __future__ import annotations

from datetime import UTC, datetime, timedelta
from typing import Any, cast

from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
from jose import JWTError, jwt
from passlib.context import CryptContext
from sqlalchemy import select
from sqlalchemy.ext.asyncio import AsyncSession

from app.core.config import get_settings
from app.db.models import User
from app.db.session import get_db

pwd_context = CryptContext(schemes=["pbkdf2_sha256"], deprecated="auto")
bearer_scheme = HTTPBearer(auto_error=False)


def hash_password(raw_password: str) -> str:
    return cast(str, pwd_context.hash(raw_password))


def verify_password(raw_password: str, hashed_password: str) -> bool:
    return cast(bool, pwd_context.verify(raw_password, hashed_password))


def create_access_token(subject: str, email: str) -> tuple[str, int]:
    settings = get_settings()
    expires_delta = timedelta(minutes=settings.jwt_access_token_expire_minutes)
    expires_at = datetime.now(UTC) + expires_delta
    payload = {"sub": subject, "email": email, "exp": expires_at}
    token = jwt.encode(payload, settings.jwt_secret_key, algorithm=settings.jwt_algorithm)
    return token, int(expires_delta.total_seconds())


def decode_access_token(token: str) -> dict[str, Any]:
    settings = get_settings()
    decoded = jwt.decode(token, settings.jwt_secret_key, algorithms=[settings.jwt_algorithm])
    return cast(dict[str, Any], decoded)


def _unauthorized(detail: str = "Authentication required.") -> HTTPException:
    return HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail={"code": "UNAUTHORIZED", "message": detail},
    )


async def get_current_user(
    credentials: HTTPAuthorizationCredentials | None = Depends(bearer_scheme),
    db: AsyncSession = Depends(get_db),
) -> User:
    if credentials is None:
        raise _unauthorized()

    try:
        payload = decode_access_token(credentials.credentials)
    except JWTError as exc:
        raise _unauthorized("Invalid or expired token.") from exc

    user_id = payload.get("sub")
    if not user_id:
        raise _unauthorized("Token missing subject.")

    stmt = select(User).where(User.id == user_id, User.is_active.is_(True))
    user = (await db.execute(stmt)).scalar_one_or_none()
    if user is None:
        raise _unauthorized("User not found.")
    return user
